requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.26
gunicorn==20.1.0  
python-3.8.0
